refactor: extract loan functionality to mixin
